requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2.8
django-crispy-forms==1.13.0
djangorestframework==3.13.1
mysqlclient==2.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
